Separate leaflet map to its own component
